requirements.txt
psutil>=5.9.0
pynput>=1.7.6
plyer>=2.1.0
pywin32>=305
pystray>=0.19.0
pillow>=10.0.0
types-pywin32>=305.0.0